wedding_invitation_system/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
qrcode==7.4.2
Pillow==10.0.1
requests==2.31.0
python-dotenv==1.0.0
selenium==4.21.0
webdriver-manager==4.0.2
pytz==2023.3
pandas>=2.1.1
openpyxl>=3.1.2